implementacion de modulos

+++ src/main/java/com/pharmman/backend/dto/request/CrearUsuarioRequest.java
@@ -9,5 +9,5 @@ public class CrearUsuarioRequest {
     private String apellidoMaterno;
     private String email;
     private String password;
-    private Long rolId;
+    private Integer rolId;
 }
