+++ src/main/java/com/pharmman/backend/repository/IRolRepository.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.pharmman.backend.entity.Rol;
 
 @Repository
-public interface IRolRepository extends JpaRepository<Rol, Long> {
+public interface IRolRepository extends JpaRepository<Rol, Integer> {
     Optional<Rol> findByNombre(String nombre);
+    
 }
